LayoutTests/js/dom/modules/module-inline-current-script.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test document.currentScript should be null when executing the inlined module.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script>
debug('Module is not executed yet.');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module">

debug("Module execution is confined in the module environment.");
shouldBe("document.currentScript", "null");

finishJSTest();
</script>
</body>
</html>
